update functions/dcupdate.fish: now only displaying current version if no args are given

+++ functions/dcupdate.fish
@@ -1,8 +1,12 @@
 function dcupdate
-  echo "--- closing remaining discord instances"
-  killall discord > /dev/null
-  killall Discord > /dev/null
-
-  echo "running: \$ sudo sed -i \"s![0-9]\{1,\}[.][0-9]\{1,\}[.][0-9]\{1,\}!$argv!g\" /opt/discord/resources/build_info.json"
-  sudo sed -i "s![0-9]\{1,\}[.][0-9]\{1,\}[.][0-9]\{1,\}!$argv!g" /opt/discord/resources/build_info.json
+  if not set -q argv[1]
+    echo "current version: $(grep -Po "[0-9]{1,}[.][0-9]{1,}[.][0-9]{1,}" /opt/discord/resources/build_info.json)" 
+  else
+    echo "--- closing remaining discord instances"
+    killall discord > /dev/null
+    killall Discord > /dev/null
+  
+    echo "running: \$ sudo sed -i \"s![0-9]\{1,\}[.][0-9]\{1,\}[.][0-9]\{1,\}!$argv!g\" /opt/discord/resources/build_info.json"
+    sudo sed -i "s![0-9]\{1,\}[.][0-9]\{1,\}[.][0-9]\{1,\}!$argv!g" /opt/discord/resources/build_info.json
+  end
 end
